Add files via upload
More introductory java codes.
Usernames: code that taught me how to work with system disk, using the File and FIleNotFoundException and so on
Vers: a more versatile billing program but does not include discounts
Year: A piece of code I'm quite proud of. It uses a switch statement to indicate how many days a chosen month has with special exception for February.
Week: uses switch statement in its basic form to output which day of the week the user has chosen.